+++ qwq/PCLCore/Minecraft/Mod/Loader/Fabric/FabricInstaller.swift
@@ -17,9 +17,9 @@ public class FabricInstaller {
     }
     
     public static func installFabric(version: MinecraftVersion, minecraftDirectory: MinecraftDirectory, runningDirectory: URL, _ loaderVersion: String) async throws {
-        let manifestURL = runningDirectory.appending(path: "\(runningDirectory.lastPathComponent).json")
+        let manifestURL = runningDirectory.appendingPathComponent("\(runningDirectory.lastPathComponent).json")
         // 若 inheritsFrom 对应的版本 JSON 不存在，复制
-        let baseManifestURL = minecraftDirectory.versionsURL.appending(path: version.displayName).appending(path: "\(version.displayName).json")
+        let baseManifestURL = minecraftDirectory.versionsURL.appendingPathComponent(version.displayName).appendingPathComponent("\(version.displayName).json")
         if !FileManager.default.fileExists(atPath: baseManifestURL.path) {
             try? FileManager.default.createDirectory(at: baseManifestURL.parent(), withIntermediateDirectories: true)
             try FileManager.default.copyItem(at: manifestURL, to: baseManifestURL)